requirements.txt
asgiref==3.8.1
sqlparse==0.5.1
Django>=4.2
djangorestframework
Pillow
djangorestframework-simplejwt
djangorestframework-simplejwt
